fix(server): Address PR review feedback
Addresses PR #1483 review comments from coderabbitai, devin-ai-integration,
and claude.

intent(observability): unify the pack_completed log schema across all
  emitters so every log-based metric label (`source`, `outcome`,
  `limitKind`) actually populates — previously `source` was only emitted
  by rateLimit.ts, and the `pack_requests` metric would have had empty
  `source` labels on 95%+ of events
decision(schema-location): extract PACK_EVENT + PackOutcome + getRepoHost
  to `actions/packEventSchema.ts` (new) and packRequestSchema to
  `actions/packRequestSchema.ts` (new). This addresses three reviewer
  findings at once: (1) PACK_EVENT duplicated across packAction and
  rateLimit, (2) PackOutcome union missing 'rate_limited', (3)
  packAction.ts exceeding the 250-line project limit (276 → 204)
decision(buildCfLogField-location): move `buildCfLogField` from
  cloudLogger.ts to logger.ts so rateLimit.ts can reuse it — rate-limit
  hits are exactly the events where Cloudflare country/ASN matters most
  for triage (bots disproportionately hit the limit). Also added early-
  return for the all-undefined case per the review suggestion
decision(validation-error-source): move `getClientInfo(c)` above the
  try/catch in packAction so the validation_error log can attach
  `source` (the other labels — inputType/repoHost/format — genuinely
  can't exist before validation, so they're only on success/pack_error)
decision(clientInfo-caveat): add explicit NOTE comment stating cf-ray /
  cf-ipcountry / cf-asn / source are spoofable and must never be used
  for auth or rate-limiting. Trust is anchored by cloudflareGuard +
  CLOUDFLARE_ORIGIN_SECRET, not these headers
rejected(getRepoHost-fallback-rename): keep the fallback default
  'github.com' for unparseable URLs. Repomix's own shorthand accepts
  'owner/repo' and resolves to github.com — the fallback matches that
  behavior. Changing to 'github.com (shorthand)' would create two
  semantically-identical buckets in the metric
