+++ website/server/src/actions/packAction.ts
@@ -1,11 +1,8 @@
 import zlib from 'node:zlib';
 import type { Context } from 'hono';
 import { stream } from 'hono/streaming';
-import { isValidRemoteValue } from 'repomix';
-import { z } from 'zod';
 import { processZipFile } from '../domains/pack/processZipFile.js';
 import { processRemoteRepo } from '../domains/pack/remoteRepo.js';
-import { FILE_SIZE_LIMITS } from '../domains/pack/utils/fileUtils.js';
 import { sanitizePattern } from '../domains/pack/utils/validation.js';
 import type { PackProgressStage, ProcessPackResult } from '../types.js';
 import { getClientInfo } from '../utils/clientInfo.js';
@@ -14,87 +11,16 @@ import { logError, logInfo } from '../utils/logger.js';
 import { calculateMemoryDiff, getMemoryUsage } from '../utils/memory.js';
 import { formatLatencyForDisplay } from '../utils/time.js';
 import { validateRequest } from '../utils/validation.js';
-
-// Unified event name for log-based metrics — every terminal pack-request log
-// carries this field so counts/distributions can be aggregated by `outcome`.
-const PACK_EVENT = 'pack_completed';
-
-type PackOutcome = 'success' | 'validation_error' | 'pack_error';
-
-// Extract a stable `repoHost` label for log-based metrics (e.g. 'github.com',
-// 'gitlab.com', 'upload'). Repomix accepts shorthand like 'owner/repo' which
-// isn't a parseable URL — default those to 'github.com' since that's the
-// shorthand target.
-function getRepoHost(input: { file?: unknown; url?: string }): string {
-  if (input.file) return 'upload';
-  const url = input.url;
-  if (!url) return 'unknown';
-  try {
-    return new URL(url).hostname;
-  } catch {
-    return 'github.com';
-  }
-}
-
-const packRequestSchema = z
-  .object({
-    url: z
-      .string()
-      .min(1, 'Repository URL is required')
-      .max(200, 'Repository URL is too long')
-      .transform((val) => val.trim())
-      .refine((val) => isValidRemoteValue(val), { message: 'Invalid repository URL' })
-      .optional(),
-    file: z
-      .custom<File>()
-      .refine((file) => file instanceof File, {
-        message: 'Invalid file format',
-      })
-      .refine((file) => file.type === 'application/zip' || file.name.endsWith('.zip'), {
-        message: 'Only ZIP files are allowed',
-      })
-      .refine((file) => file.size <= FILE_SIZE_LIMITS.MAX_ZIP_SIZE, {
-        // 10MB limit
-        message: 'File size must be less than 10MB',
-      })
-      .optional(),
-    format: z.enum(['xml', 'markdown', 'plain']),
-    options: z
-      .object({
-        removeComments: z.boolean().optional(),
-        removeEmptyLines: z.boolean().optional(),
-        showLineNumbers: z.boolean().optional(),
-        fileSummary: z.boolean().optional(),
-        directoryStructure: z.boolean().optional(),
-        includePatterns: z
-          .string()
-          .max(100_000, 'Include patterns too long')
-          .optional()
-          .transform((val) => val?.trim()),
-        ignorePatterns: z
-          .string()
-          // Regular expression to validate ignore patterns
-          // Allowed characters: alphanumeric, *, ?, /, -, _, ., !, (, ), space, comma
-          .regex(/^[a-zA-Z0-9*?/\-_.,!()\s]*$/, 'Invalid characters in ignore patterns')
-          .max(1000, 'Ignore patterns too long')
-          .optional()
-          .transform((val) => val?.trim()),
-        outputParsable: z.boolean().optional(),
-        compress: z.boolean().optional(),
-      })
-      .strict(),
-  })
-  .strict()
-  .refine((data) => data.url || data.file, {
-    message: 'Either URL or file must be provided',
-  })
-  .refine((data) => !(data.url && data.file), {
-    message: 'Cannot provide both URL and file',
-  });
+import { getRepoHost, PACK_EVENT, type PackOutcome } from './packEventSchema.js';
+import { type PackRequest, packRequestSchema } from './packRequestSchema.js';
 
 export const packAction = async (c: Context) => {
+  // Extracted up-front so validation_error logs can attach `source` for the
+  // pack_requests metric label; clientInfo doesn't depend on validated input.
+  const clientInfo = getClientInfo(c);
+
   // Parse and validate request before starting stream
-  let validatedData: z.infer<typeof packRequestSchema>;
+  let validatedData: PackRequest;
   let sanitizedOptions: { includePatterns?: string; ignorePatterns?: string } & Record<string, unknown>;
 
   try {
@@ -133,6 +59,7 @@ export const packAction = async (c: Context) => {
       event: PACK_EVENT,
       outcome: 'validation_error' satisfies PackOutcome,
       requestId: c.get('requestId'),
+      source: clientInfo.source,
     });
 
     const { handlePackError } = await import('../utils/errorHandler.js');
@@ -141,7 +68,6 @@ export const packAction = async (c: Context) => {
   }
 
   const requestId = c.get('requestId');
-  const clientInfo = getClientInfo(c);
   const inputType: 'file' | 'url' = validatedData.file ? 'file' : 'url';
   const repoHost = getRepoHost({ file: validatedData.file, url: validatedData.url });
 
@@ -233,6 +159,7 @@ export const packAction = async (c: Context) => {
         inputType,
         repoHost,
         cached,
+        source: clientInfo.source,
         durationMs: Date.now() - startTime,
         repository: result.metadata.repository,
         duration: formatLatencyForDisplay(startTime),
@@ -262,6 +189,7 @@ export const packAction = async (c: Context) => {
         format: validatedData.format,
         inputType,
         repoHost,
+        source: clientInfo.source,
       });
 
       const { handlePackError } = await import('../utils/errorHandler.js');
